Introduce IQueueMonitor and a corresponding RabbitMQ implementation of the interface
The extension method `UseQueueMonitoring` adds a Gauge that exports the number of ready messages for every queue that is consumed. Note that this mechanism is not meant for publishers right now.
Every queuing system can implement the IQueueMonitor interface and add an instance of it per queue to the DI container and the registered monitors are discovered during startup for exporting the metrics.