Rewrite connection handling for RabbitMQ
To share connections between the QueueMonitor and the consumers a new kind of ConnectionFactory is required because the previous RabbitMQConnectionFactory only created a IConnectionFactory but a IConnection.
The new implementation uses Lazy<T> to cache an IConnection and an IModel for sharing purposes.
Note that the 'CreateX' methods always create new instances if the sharing should be skipped.