Adjust order of the test and adds another timeout
in this order it is clear, that the consumption is startet before the
publishing (this is a nats special because the messages are lost when
no one is listening.
It also adds another timeout to the ConsumerWait (in the test) so we
don't wait forever for a message to come in.